refactor handlers and their tests
