+++ src/main/java/com/lilangel/teamplay/tgbot/handlers/admins/ProjectHandler.java
@@ -1,8 +1,9 @@
-package com.lilangel.teamplay.tgbot.handlers;
+package com.lilangel.teamplay.tgbot.handlers.admins;
 
 import com.lilangel.teamplay.exception.ProjectNotFoundException;
 import com.lilangel.teamplay.models.Project;
 import com.lilangel.teamplay.service.ProjectService;
+import com.lilangel.teamplay.tgbot.handlers.AbstractHandler;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
@@ -89,7 +90,6 @@ public class ProjectHandler extends AbstractHandler {
      * @param args аргументы
      * @return строка с информацией о всех проектах
      */
-    @Override
     protected String getAll(Map<String, String> args) {
         String template = """
                 \t\t\t\tID: %d
@@ -112,7 +112,6 @@ public class ProjectHandler extends AbstractHandler {
      * @param args аргументы
      * @return строка с информацией о проекте
      */
-    @Override
     protected String getById(Map<String, String> args) {
         String template = """
                 Project:
@@ -135,7 +134,6 @@ public class ProjectHandler extends AbstractHandler {
      * @param args аргументы
      * @return строка с идентификатором созданного проекта
      */
-    @Override
     protected String create(Map<String, String> args) {
         if (args.size() != ARGS_COUNT_TO_CREATE) {
             return "Wrong args number";
@@ -155,7 +153,6 @@ public class ProjectHandler extends AbstractHandler {
      * @param args аргументы
      * @return строка с результатом
      */
-    @Override
     protected String deleteById(Map<String, String> args) {
         try {
             projectService.deleteById(Integer.parseInt(args.get("id")));
