+++ src/main/java/com/lilangel/teamplay/tgbot/handlers/admins/TeamHandler.java
@@ -1,8 +1,9 @@
-package com.lilangel.teamplay.tgbot.handlers;
+package com.lilangel.teamplay.tgbot.handlers.admins;
 
 import com.lilangel.teamplay.exception.TeamNotFoundException;
 import com.lilangel.teamplay.models.Team;
 import com.lilangel.teamplay.service.TeamService;
+import com.lilangel.teamplay.tgbot.handlers.AbstractHandler;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
@@ -89,7 +90,6 @@ public class TeamHandler extends AbstractHandler {
      * @param args аргументы
      * @return строка с информацией о всех проектах
      */
-    @Override
     protected String getAll(Map<String, String> args) {
         String template = """
                 \t\t\t\tID: %d
@@ -111,7 +111,6 @@ public class TeamHandler extends AbstractHandler {
      * @param args аргументы
      * @return строка с информацией о команде
      */
-    @Override
     protected String getById(Map<String, String> args) {
         String template = """
                 Team:
@@ -133,7 +132,6 @@ public class TeamHandler extends AbstractHandler {
      * @param args аргументы
      * @return строка с идентификатором созданной команды
      */
-    @Override
     protected String create(Map<String, String> args) {
         if (args.size() != ARGS_COUNT_TO_CREATE) {
             return "Wrong args number";
@@ -152,7 +150,6 @@ public class TeamHandler extends AbstractHandler {
      * @param args аргументы
      * @return строка с результатом
      */
-    @Override
     protected String deleteById(Map<String, String> args) {
         try {
             teamService.deleteById(Integer.parseInt(args.get("id")));
