+++ src/test/java/com/lilangel/teamplay/tgbot/handlers/admins/EmployerHandlerTest.java
@@ -1,4 +1,4 @@
-package com.lilangel.teamplay.tgbot.handlers;
+package com.lilangel.teamplay.tgbot.handlers.admins;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
